build: update .gitignore and add new files for improved structure generation functionality
