+++ pyproject.toml
@@ -15,6 +15,7 @@ dependencies = [
     "litellm>=1.76.1",
     "google-adk==1.14.0",
     "pre-commit>=4.3.0",
+    "bohrium-sdk>=0.14.0",
 ]
 
 [tool.setuptools]
